moved all typescript classes into different folder so its neater
