+++ src/lib/classes/IncrementalNumberInput.ts
