+++ src/lib/classes/Input.ts
