+++ src/lib/classes/NumberInput.ts
