+++ src/lib/classes/ToggleInput.ts
